% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_aoi_to_chm.R
\name{combine_rvb_irc}
\alias{combine_rvb_irc}
\title{Combiner les ortho RVB et IRC en image 4 bandes (R, G, B, PIR)}
\usage{
combine_rvb_irc(rvb, irc)
}
\arguments{
\item{rvb}{SpatRaster ortho RVB (3 bandes : Rouge, Vert, Bleu)}

\item{irc}{SpatRaster ortho IRC (3 bandes : PIR, Rouge, Vert)}
}
\value{
SpatRaster 4 bandes (Rouge, Vert, Bleu, PIR)
}
\description{
Le modèle Open-Canopy attend 4 canaux : Rouge, Vert, Bleu, PIR
\itemize{
\item RVB fournit les 3 premières bandes
\item IRC fournit le PIR (bande 1 de l'IRC = Proche Infrarouge)
}
}
\keyword{internal}
