% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{compute_canopy_change}
\alias{compute_canopy_change}
\title{Différence de canopée entre deux dates}
\usage{
compute_canopy_change(chm_t1, chm_t2)
}
\description{
Différence de canopée entre deux dates
}
