% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_open_canopy.R
\name{compute_forest_cover_grid}
\alias{compute_forest_cover_grid}
\title{Couverture forestière par grille}
\usage{
compute_forest_cover_grid(chm_raster, cell_size = 100, height_threshold = 2)
}
\arguments{
\item{chm_raster}{SpatRaster du CHM}

\item{cell_size}{Côté de la maille en mètres}

\item{height_threshold}{Hauteur minimale, en mètres, pour compter une cellule
comme forestière}
}
\value{
SpatRaster de la fraction forestière par maille, dans \verb{[0, 1]}
}
\description{
Couverture forestière par grille
}
