% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{compute_gndvi}
\alias{compute_gndvi}
\title{Calculer le GNDVI (Green NDVI) depuis une ortho IRC IGN}
\usage{
compute_gndvi(irc_raster)
}
\arguments{
\item{irc_raster}{SpatRaster IRC IGN}
}
\value{
SpatRaster du GNDVI
}
\description{
GNDVI = (PIR - Vert) / (PIR + Vert)
}
