% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{compute_irc_stats}
\alias{compute_irc_stats}
\title{Statistiques spectrales d'une ortho IRC IGN}
\usage{
compute_irc_stats(irc_raster, max_cells = 1e+06)
}
\arguments{
\item{irc_raster}{SpatRaster IRC IGN}

\item{max_cells}{Taille max de l'échantillon régulier (médiane et parts)}
}
\value{
data.frame avec statistiques par bande + NDVI
}
\description{
Statistiques spectrales d'une ortho IRC IGN
}
