% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{compute_ndvi}
\alias{compute_ndvi}
\title{Calculer le NDVI depuis une ortho IRC IGN}
\usage{
compute_ndvi(irc_raster)
}
\arguments{
\item{irc_raster}{SpatRaster IRC IGN (bandes: PIR, Rouge, Vert)}
}
\value{
SpatRaster du NDVI (valeurs entre -1 et 1)
}
\description{
NDVI = (PIR - Rouge) / (PIR + Rouge)
Nécessite une image IRC avec bandes PIR et Rouge
}
