% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{compute_ndwi}
\alias{compute_ndwi}
\title{Calculer le NDWI (Normalized Difference Water Index, McFeeters 1996)}
\usage{
compute_ndwi(irc_raster)
}
\arguments{
\item{irc_raster}{SpatRaster IRC IGN (bandes: PIR, Rouge, Vert)}
}
\value{
SpatRaster du NDWI (valeurs entre -1 et 1)
}
\description{
NDWI = (Vert - PIR) / (Vert + PIR)
Valeurs > 0 typiquement = eau ; utile pour masquer rivières/étangs du CHM.
}
