% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{compute_savi}
\alias{compute_savi}
\title{Calculer le NDRE (Normalized Difference Red Edge) approximé}
\usage{
compute_savi(irc_raster, L = 0.5)
}
\arguments{
\item{irc_raster}{SpatRaster IRC IGN}
}
\value{
SpatRaster
}
\description{
Approximation basée sur la différence PIR - Vert normalisée.
Utile pour différencier les espèces et l'état de santé.
}
