% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{cross_ndvi_chm}
\alias{cross_ndvi_chm}
\title{Croiser NDVI (IGN IRC) et CHM pour caractériser la végétation}
\usage{
cross_ndvi_chm(ndvi_raster, chm_raster)
}
\arguments{
\item{ndvi_raster}{SpatRaster NDVI (depuis ortho IRC IGN)}

\item{chm_raster}{SpatRaster CHM LiDAR}
}
\value{
SpatRaster avec classes croisées
}
\description{
Croiser NDVI (IGN IRC) et CHM pour caractériser la végétation
}
