% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_open_canopy.R
\name{detect_canopy_loss}
\alias{detect_canopy_loss}
\title{Détection de perte de canopée}
\usage{
detect_canopy_loss(chm_t1, chm_t2, threshold = -5)
}
\description{
Détection de perte de canopée
}
