% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_aoi_to_chm.R
\name{.repertoire_python_embarque}
\alias{.repertoire_python_embarque}
\title{Répertoire du module Python embarqué (timmnet_standalone)}
\usage{
.repertoire_python_embarque()
}
\value{
Chemin du dossier \code{inst/python} du package
}
\description{
Répertoire du module Python embarqué (timmnet_standalone)
}
\keyword{internal}
