% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_open_canopy.R
\name{download_ign_ortho_pair}
\alias{download_ign_ortho_pair}
\title{Télécharger les ortho RVB et IRC pour une même emprise}
\usage{
download_ign_ortho_pair(
  bbox,
  res_m = RES_IGN,
  millesime_ortho = MILLESIME_ORTHO,
  millesime_irc = MILLESIME_IRC
)
}
\arguments{
\item{bbox}{Emprise c(xmin, ymin, xmax, ymax) en Lambert-93 (EPSG:2154)}

\item{res_m}{Résolution en mètres}

\item{millesime_ortho}{Millésime ortho RVB (NULL = plus récent)}

\item{millesime_irc}{Millésime IRC (NULL = plus récent)}
}
\value{
Liste avec les chemins des fichiers RVB et IRC et millésimes
}
\description{
Télécharger les ortho RVB et IRC pour une même emprise
}
