% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_open_canopy.R
\name{download_ign_wms}
\alias{download_ign_wms}
\title{Télécharger une ortho IGN via WMS (pour une emprise donnée)}
\usage{
download_ign_wms(
  bbox,
  layer = IGN_LAYER_ORTHO,
  res_m = RES_IGN,
  dest_file = NULL
)
}
\arguments{
\item{bbox}{Emprise c(xmin, ymin, xmax, ymax) en Lambert-93 (EPSG:2154)}

\item{layer}{Couche WMS (RVB ou IRC)}

\item{res_m}{Résolution en mètres (0.2 par défaut)}

\item{dest_file}{Chemin du fichier de sortie}
}
\value{
Chemin local du fichier téléchargé
}
\description{
Télécharger une ortho IGN via WMS (pour une emprise donnée)
}
