% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_open_canopy.R
\name{download_open_canopy_subset}
\alias{download_open_canopy_subset}
\title{Télécharger un sous-ensemble du dataset Open-Canopy}
\usage{
download_open_canopy_subset(
  split = "test",
  n_tiles = 5,
  data_type = "all",
  dest_dir = DATA_DIR_HF,
  token = HF_TOKEN
)
}
\arguments{
\item{split}{"train", "val", ou "test"}

\item{n_tiles}{Nombre de tuiles}

\item{data_type}{"images" (SPOT), "lidar" (CHM), "lidar_v2", ou "all"}

\item{dest_dir}{Répertoire de destination}

\item{token}{Token Hugging Face}
}
\description{
Télécharger un sous-ensemble du dataset Open-Canopy
}
