% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_open_canopy.R
\name{hf_download_file}
\alias{hf_download_file}
\title{Télécharger un fichier depuis Hugging Face}
\usage{
hf_download_file(repo_id, filename, dest_dir, token = "", overwrite = FALSE)
}
\arguments{
\item{repo_id}{Identifiant du dépôt}

\item{filename}{Chemin du fichier dans le dépôt}

\item{dest_dir}{Répertoire de destination local}

\item{token}{Token Hugging Face (optionnel)}

\item{overwrite}{Écraser si le fichier existe déjà}
}
\value{
Chemin local du fichier téléchargé
}
\description{
Télécharger un fichier depuis Hugging Face
}
