% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_open_canopy.R
\name{hf_download_files}
\alias{hf_download_files}
\title{Télécharger un ensemble de fichiers depuis Hugging Face}
\usage{
hf_download_files(repo_id, file_list, dest_dir, token = "", overwrite = FALSE)
}
\description{
Télécharger un ensemble de fichiers depuis Hugging Face
}
