% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_open_canopy.R
\name{hf_list_files}
\alias{hf_list_files}
\title{Lister les fichiers du dataset sur Hugging Face}
\usage{
hf_list_files(repo_id, path = "", token = "")
}
\arguments{
\item{repo_id}{Identifiant du dépôt (ex: "AI4Forest/Open-Canopy")}

\item{path}{Chemin dans le dépôt (ex: "" pour la racine)}

\item{token}{Token Hugging Face (optionnel)}
}
\value{
data.frame avec les informations des fichiers
}
\description{
Lister les fichiers du dataset sur Hugging Face
}
