% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_open_canopy.R,
%   R/pipeline_aoi_to_chm.R
\name{ign_layer_name}
\alias{ign_layer_name}
\title{Construire le nom de couche WMS en fonction du millésime}
\usage{
ign_layer_name(type = c("ortho", "irc"), millesime = NULL)

ign_layer_name(type = c("ortho", "irc"), millesime = NULL)
}
\arguments{
\item{type}{"ortho" ou "irc"}

\item{millesime}{Année (chaîne ou numérique), NULL pour la couche courante}
}
\value{
Nom de couche WMS IGN

Nom de couche WMS IGN
}
\description{
Construire le nom de couche WMS en fonction du millésime

Construire le nom de couche WMS en fonction du millésime
}
\keyword{internal}
