% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_aoi_to_chm.R
\name{load_aoi}
\alias{load_aoi}
\title{Charger l'AOI depuis un fichier GeoPackage}
\usage{
load_aoi(gpkg_path, layer = NULL)
}
\arguments{
\item{gpkg_path}{Chemin vers le fichier .gpkg}

\item{layer}{Nom de la couche (NULL = première couche)}
}
\value{
sf object en Lambert-93 (EPSG:2154)
}
\description{
Charger l'AOI depuis un fichier GeoPackage
}
