% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{load_chm}
\alias{load_chm}
\title{Charger un CHM dérivé du LiDAR}
\usage{
load_chm(tile_path)
}
\arguments{
\item{tile_path}{Chemin vers le fichier .tif}
}
\value{
SpatRaster
}
\description{
Charger un CHM dérivé du LiDAR
}
\keyword{internal}
