% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_open_canopy.R
\name{load_predictions}
\alias{load_predictions}
\title{Charger les prédictions}
\usage{
load_predictions(prediction_path)
}
\arguments{
\item{prediction_path}{Chemin d'un fichier de prédiction, ou d'un répertoire
dont les .tif seront tous chargés}
}
\value{
SpatRaster, ou liste de SpatRaster si \code{prediction_path} est un dossier
}
\description{
Charger les prédictions
}
