% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{load_spot_image}
\alias{load_spot_image}
\title{Charger une image SPOT 6-7 (Open-Canopy)}
\usage{
load_spot_image(tile_path)
}
\arguments{
\item{tile_path}{Chemin vers le fichier .tif}
}
\value{
SpatRaster
}
\description{
Charger une image SPOT 6-7 (Open-Canopy)
}
\keyword{internal}
