% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{load_tiles}
\alias{load_tiles}
\title{Lister et charger toutes les tuiles d'un split}
\usage{
load_tiles(split = "test", data_type = "images", data_dir = DATA_DIR_HF)
}
\description{
Lister et charger toutes les tuiles d'un split
}
\keyword{internal}
