% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{mask_vegetation}
\alias{mask_vegetation}
\title{Créer un masque de végétation à partir du NDVI}
\usage{
mask_vegetation(ndvi_raster, threshold = 0.3)
}
\arguments{
\item{ndvi_raster}{SpatRaster du NDVI}

\item{threshold}{Seuil NDVI pour considérer de la végétation}
}
\value{
SpatRaster binaire (1 = végétation)
}
\description{
Créer un masque de végétation à partir du NDVI
}
