% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_open_canopy.R
\name{normalize_for_model}
\alias{normalize_for_model}
\title{Préparer une tuile IGN pour le modèle Open-Canopy}
\usage{
normalize_for_model(tile, normalize_to = "0_1")
}
\arguments{
\item{tile}{SpatRaster d'une tuile}

\item{normalize_to}{Plage cible ("0_1" ou "0_10000")}
}
\value{
SpatRaster normalisé
}
\description{
Normalisation des valeurs et conversion en format attendu.
Les images SPOT Open-Canopy sont en réflectance \verb{[0, 1]} ou \verb{[0, 10000]}.
Les ortho IGN sont en radiométrie 8-bit \verb{[0, 255]}.
}
