% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{plot_canopy_change}
\alias{plot_canopy_change}
\title{Visualiser les changements de canopée}
\usage{
plot_canopy_change(
  change_raster,
  title = "Changement de hauteur de canopée (m)"
)
}
\arguments{
\item{change_raster}{SpatRaster de variation de hauteur, tel que renvoyé par
\code{\link[=compute_canopy_change]{compute_canopy_change()}}}

\item{title}{Titre du graphique}
}
\value{
Invisible \code{NULL}, appelée pour son effet de bord (tracé)
}
\description{
Visualiser les changements de canopée
}
