% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{plot_canopy_classes}
\alias{plot_canopy_classes}
\title{Classification de la canopée}
\usage{
plot_canopy_classes(
  chm_raster,
  breaks = c(0, 2, 5, 10, 20, Inf),
  labels = c("Sol/herbe (<2m)", "Arbustes (2-5m)", "Petits arbres (5-10m)",
    "Arbres moyens (10-20m)", "Grands arbres (>20m)"),
  title = "Classes de hauteur de canopée"
)
}
\description{
Classification de la canopée
}
