% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{plot_ign_full_comparison}
\alias{plot_ign_full_comparison}
\title{Comparaison multi-panneaux : IGN RVB + IRC + NDVI + CHM}
\usage{
plot_ign_full_comparison(
  rvb_raster,
  irc_raster,
  chm_raster = NULL,
  tile_name = ""
)
}
\arguments{
\item{rvb_raster}{SpatRaster ortho RVB IGN}

\item{irc_raster}{SpatRaster ortho IRC IGN}

\item{chm_raster}{SpatRaster CHM (optionnel)}

\item{tile_name}{Nom de la tuile}
}
\description{
Comparaison multi-panneaux : IGN RVB + IRC + NDVI + CHM
}
