% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{plot_irc}
\alias{plot_irc}
\title{Visualiser une ortho IRC en fausses couleurs}
\usage{
plot_irc(irc_raster, title = "Ortho IRC IGN (0.20m)")
}
\arguments{
\item{irc_raster}{SpatRaster IRC IGN (PIR, Rouge, Vert)}

\item{title}{Titre}
}
\description{
Affichage IRC classique : PIR en rouge, Rouge en vert, Vert en bleu
La végétation active apparaît en rouge vif.
}
