% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{plot_ndvi}
\alias{plot_ndvi}
\title{Visualiser le NDVI}
\usage{
plot_ndvi(ndvi_raster, title = "NDVI")
}
\arguments{
\item{ndvi_raster}{SpatRaster du NDVI}

\item{title}{Titre}
}
\description{
Visualiser le NDVI
}
