% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{plot_resolution_comparison}
\alias{plot_resolution_comparison}
\title{Comparaison SPOT (1.5m) vs IGN (0.20m) sur la même zone}
\usage{
plot_resolution_comparison(spot_raster, ign_raster, tile_name = "")
}
\arguments{
\item{spot_raster}{SpatRaster SPOT}

\item{ign_raster}{SpatRaster ortho IGN RVB}

\item{tile_name}{Nom de la tuile}
}
\description{
Comparaison SPOT (1.5m) vs IGN (0.20m) sur la même zone
}
