% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{plot_rgb}
\alias{plot_rgb}
\title{Visualiser une image en couleurs naturelles (RGB)}
\usage{
plot_rgb(raster_rgb, title = "Image RGB", bands = c(1, 2, 3))
}
\arguments{
\item{raster_rgb}{SpatRaster avec au moins 3 bandes}

\item{title}{Titre du graphique}

\item{bands}{Indices des bandes RGB}
}
\description{
Fonctionne avec SPOT (Open-Canopy) et ortho IGN RVB
}
