% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_open_canopy.R
\name{predict_chm_from_ign}
\alias{predict_chm_from_ign}
\title{Pipeline complet : ortho IGN → prédiction CHM}
\usage{
predict_chm_from_ign(ign_path, model_path, ign_type = "rvb")
}
\arguments{
\item{ign_path}{Chemin de l'ortho IGN (.jp2 ou .tif)}

\item{model_path}{Chemin du modèle pré-entraîné}

\item{ign_type}{"rvb" ou "irc"}
}
\value{
SpatRaster des prédictions mosaïquées
}
\description{
Pipeline complet : ortho IGN → prédiction CHM
}
