% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_aoi_to_chm.R
\name{resample_to_spot}
\alias{resample_to_spot}
\title{Agréger une ortho IGN de 0.20m vers 1.5m}
\usage{
resample_to_spot(ign_raster)
}
\arguments{
\item{ign_raster}{SpatRaster à 0.20m}
}
\value{
SpatRaster à ~1.5m
}
\description{
Agréger une ortho IGN de 0.20m vers 1.5m
}
\keyword{internal}
