% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_open_canopy.R
\name{run_inference_python}
\alias{run_inference_python}
\title{Exécuter l'inférence sur une tuile via Python}
\usage{
run_inference_python(tile_path, model_path, output_path = NULL)
}
\arguments{
\item{tile_path}{Chemin du fichier raster de la tuile}

\item{model_path}{Chemin du modèle .ckpt}

\item{output_path}{Chemin de sortie}
}
\value{
Chemin du fichier de prédiction
}
\description{
Exécuter l'inférence sur une tuile via Python
}
