% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_open_canopy.R
\name{scan_ign_files}
\alias{scan_ign_files}
\title{Scanner un répertoire pour trouver les ortho IGN}
\usage{
scan_ign_files(dir_path = DATA_DIR_IGN, recursive = TRUE)
}
\arguments{
\item{dir_path}{Répertoire à scanner}

\item{recursive}{Recherche récursive}
}
\value{
data.frame avec les fichiers trouvés et leur type
}
\description{
Scanner un répertoire pour trouver les ortho IGN
}
