% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_open_canopy.R
\name{tile_for_inference}
\alias{tile_for_inference}
\title{Découper une image IGN en tuiles pour l'inférence}
\usage{
tile_for_inference(ign_raster, tile_size = 1000, overlap = 0)
}
\arguments{
\item{ign_raster}{SpatRaster IGN (déjà rééchantillonné à 1.5m)}

\item{tile_size}{Taille des tuiles en mètres (1000 = 1km)}

\item{overlap}{Chevauchement entre tuiles en mètres}
}
\value{
Liste de SpatRasters (tuiles)
}
\description{
Les modèles Open-Canopy travaillent sur des tuiles de 1km x 1km.
À 1.5m de résolution : 667 x 667 pixels.
}
