% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_open_canopy.R
\name{upsample_chm_to_ign}
\alias{upsample_chm_to_ign}
\title{Suréchantillonner un CHM prédit (1.5m) vers la résolution IGN (0.20m)}
\usage{
upsample_chm_to_ign(chm_predicted, target_res = RES_IGN, method = "bilinear")
}
\arguments{
\item{chm_predicted}{SpatRaster CHM prédit à 1.5m}

\item{target_res}{Résolution cible (0.2m)}

\item{method}{Méthode d'interpolation}
}
\value{
SpatRaster à 0.20m
}
\description{
Suréchantillonner un CHM prédit (1.5m) vers la résolution IGN (0.20m)
}
