% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_open_canopy.R
\name{zonal_canopy_stats}
\alias{zonal_canopy_stats}
\title{Statistiques zonales}
\usage{
zonal_canopy_stats(chm_raster, zones, fun = "mean")
}
\arguments{
\item{chm_raster}{SpatRaster du CHM}

\item{zones}{SpatVector des zones d'agrégation}

\item{fun}{Fonction d'agrégation passée à \code{\link[terra:zonal]{terra::zonal()}} (ex. "mean")}
}
\value{
data.frame d'une ligne par zone
}
\description{
Statistiques zonales
}
