% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_open_canopy.R
\name{compute_forest_cover_grid}
\alias{compute_forest_cover_grid}
\title{Couverture forestière par grille}
\usage{
compute_forest_cover_grid(chm_raster, cell_size = 100, height_threshold = 2)
}
\description{
Couverture forestière par grille
}
