% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_open_canopy.R
\name{hf_git_clone}
\alias{hf_git_clone}
\title{Cloner le dataset complet via git (nécessite git-lfs)}
\usage{
hf_git_clone(
  dest_dir = file.path(DATA_DIR_HF, "Open-Canopy"),
  token = HF_TOKEN
)
}
\description{
Cloner le dataset complet via git (nécessite git-lfs)
}
