% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{plot_canopy_change}
\alias{plot_canopy_change}
\title{Visualiser les changements de canopée}
\usage{
plot_canopy_change(
  change_raster,
  title = "Changement de hauteur de canopée (m)"
)
}
\description{
Visualiser les changements de canopée
}
