% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_open_canopy.R
\name{plot_chm_histogram}
\alias{plot_chm_histogram}
\title{Histogramme des hauteurs de canopée}
\usage{
plot_chm_histogram(
  chm_raster,
  title = "Distribution des hauteurs de canopée",
  n_breaks = 50,
  max_cells = 1e+06
)
}
\description{
Histogramme des hauteurs de canopée
}
