% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_open_canopy.R
\name{plot_prediction_comparison}
\alias{plot_prediction_comparison}
\title{Visualiser la comparaison prédiction vs référence}
\usage{
plot_prediction_comparison(
  prediction,
  reference,
  metrics = NULL,
  tile_name = "",
  n_points = 50000
)
}
\description{
Visualiser la comparaison prédiction vs référence
}
